backend/requirements.txt
fastapi==0.111.*
python-snap7==1.2.*
uvicorn[standard]==0.29.*
setuptools